Organized GUI classes for better encapsulation
