+++ source/gui/ControlPanelWidget.cpp
@@ -1,0 +1,42 @@
+#include <QHBoxLayout>
+
+#include "ControlPanelWidget.hpp"
+
+ControlPanelWidget::ControlPanelWidget(QWidget* parent):
+    QWidget(parent),
+    connection { new QPushButton("Connect") },
+    execution { new QPushButton("Start/Stop") }
+{
+    execution->setEnabled(false);
+
+    QHBoxLayout* layout = new QHBoxLayout;
+
+    layout->addWidget(connection);
+    layout->addWidget(execution);
+    layout->setAlignment(Qt::AlignCenter);
+    setLayout(layout);
+}
+
+void ControlPanelWidget::connectionChanged(bool status)
+{
+    if (status)
+    {
+        connection->setText("Disconnect");
+        execution->setText("Start");
+        execution->setEnabled(true);
+    }
+    else
+    {
+        connection->setText("Connect");
+        execution->setText("Start/Stop");
+        execution->setEnabled(false);
+    }
+}
+
+void ControlPanelWidget::executionChanged(bool status)
+{
+    if (status)
+        execution->setText("Stop");
+    else
+        execution->setText("Start");
+}
