+++ source/gui/MainWindow.hpp
@@ -1,47 +1,16 @@
-// The Application looks as the following:
-// +------------------------------------------------------------------------------------------------------------------+
-// |    CH0 - Data                                                CH1 - Data                                          |
-// |    +-------+----------+--------------+-------------+         +-------+----------+--------------+-------------+   |
-// |    | Index | Barcode  | Current [nA] | Voltage [V] |         | Index | Barcode  | Current [nA] | Voltage [V] |   |
-// |    |-------|----------|--------------|-------------|         |-------|----------|--------------|-------------|   |
-// |    | 0     | MSU00000 | 0.00         | 0.00        |         | 0     | MSU00000 | 0.00         | 0.00        |   |
-// |    | 1     | MSU00000 | 0.00         | 0.00        |         | 1     | MSU00000 | 0.00         | 0.00        |   |
-// |    | 2     | MSU00000 | 0.00         | 0.00        |         | 2     | MSU00000 | 0.00         | 0.00        |   |
-// |    +-------+----------+--------------+-------------+         +-------+----------+--------------+-------------+   |
-// |                                                                                                                  |
-// |    CH0 - Status                                              CH1 - Status                                        |
-// |    +-----------------------------------------------+         +-----------------------------------------------+   |
-// |    |                                               |         |                                               |   |
-// |    +-----------------------------------------------+         +-----------------------------------------------+   |
-// |                                                                                                                  |
-// |                                                    Test Status                                                   |
-// |                                           +---------------------------+                                          |
-// |                                           |  Time Started:    00:00   |                                          |
-// |                                           |  Time Elapsed:    0 min   |                                          |
-// |                                           |  Time Remaining:  0 min   |                                          |
-// |                                           +---------------------------+                                          |
-// +------------------------------------------------------------------------------------------------------------------+
-
 #pragma once
 
 #include <string>
-#include <memory>
-#include <utility>
+#include <vector>
 
-#include <QMenu>
-#include <QLabel>
-#include <QFrame>
-#include <QAction>
-#include <QGroupBox>
-#include <QTableView>
+#include <QWidget>
 #include <QMainWindow>
-#include <QActionGroup>
-
-#include <spdlog/spdlog.h>
 
 #include "TestInfo.hpp"
+#include "ChannelWidget.hpp"
 #include "TestController.hpp"
-#include "CollectionModel.hpp"
+#include "TestStatusWidget.hpp"
+#include "ControlPanelWidget.hpp"
 
 class MainWindow : public QMainWindow
 {
@@ -51,70 +20,25 @@ public:
     explicit MainWindow(QWidget* parent = nullptr);
     ~MainWindow();
 
+    void readInTestSettings(std::string path);
+
+public slots:
+    void receiveRequestToConnect();
+    void receiveRequestToDisconnect();
+    void receiveRequestToStart();
+    void receiveRequestToStop();
+
+    void alertUser();
+
 signals:
-    void connect(msu_smdt::Port port);
-    void startTest(std::vector<int> channels);
-    void distributeTestConfiguration(TestConfiguration normalConfig, TestConfiguration reverseConfig);
-    void fillFakeBarcodes();
-
-private slots:
-    void hasConnected();
-    void hasDisconnected();
-
-    void start();
-    void getPortInfoPSU();
-    void openHelp();
-    void openGitHub();
-    void about();
-
-    void updateTime(std::string elapsedTime, std::string remainingTime);
-
-    void alert(std::string message);
-
-    void readConfigurationFile();
 
 private:
-    msu_smdt::Port PSUPort;
-    msu_smdt::Port HWPort;
+    // TestController* controller;
 
-    TestParameters parameters;
-    TestController* controller;
+    QWidget* channelWidgetContainer;
 
-    CollectionModel* CH0Model;
-    CollectionModel* CH1Model;
-
-    QTableView* CH0View;
-    QTableView* CH1View;
-
-    QLabel* CH0Status;
-    QLabel* CH1Status;
-    QLabel* connectionStatus;
-    QLabel* timeStarted;
-    QLabel* timeElapsed;
-    QLabel* timeRemaining;
-
-    QFrame* channelFrame;
-    QFrame* testStatusFrame;
-
-    QGroupBox* CH0DataBox;
-    QGroupBox* CH1DataBox;
-    QGroupBox* CH0StatusBox;
-    QGroupBox* CH1StatusBox;
-    QGroupBox* testStatusBox;
-
-    QMenu* connectMenu;
-    QMenu* testMenu;
-    QMenu* helpMenu;
-
-    QAction* connectAct;
-    QAction* disconnectAct;
-    QAction* startAct;
-    QAction* stopAct;
-    QAction* openHelpAct;
-    QAction* openGitHubAct;
-    QAction* aboutAct;
-
-    QAction* createFakes;
-
-    std::shared_ptr<spdlog::logger> logger;
+    ControlPanelWidget* controlPanelWidget;
+    ChannelWidget* channelWidgetLeft;
+    ChannelWidget* channelWidgetRight;
+    TestStatusWidget* testStatusWidget;
 };
