+++ source/main.cpp
@@ -19,10 +19,21 @@
 #include <spdlog/spdlog.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
 
+#define TEST_GUI
+#ifdef TEST_GUI
+    #include <QWidget>
+    #include <QVBoxLayout>
+
+    #include <gui/ChannelWidget.hpp>
+    #include <gui/TestStatusWidget.hpp>
+    #include <gui/ControlPanelWidget.hpp>
+    #include <gui/TestInfo.hpp>
+#endif
+
+#include <gui/MainWindow.hpp>
+
 #ifdef NO_GUI
     #include "test/manual/test.hpp"
-#else
-    #include <gui/MainWindow.hpp>
 #endif 
 
 int main(int argc, char** argv)
@@ -31,12 +42,26 @@ int main(int argc, char** argv)
     spdlog::set_level(spdlog::level::info);
     spdlog::set_pattern("[%Y-%m-%d %H:%M:%S] [%=16n] %^[%=8l]%$ %v");
 
-#ifndef NO_GUI
+/*
+#ifdef TEST_GUI
+    TestParameters params;
+
     QApplication app(argc, argv);
     app.setStyle(QStyleFactory::create("Fusion"));
 
-    MainWindow main_window;
-    main_window.show();
+    QWidget mainWidget;
+    ChannelWidget channelWidget(nullptr, 0, params);
+    TestStatusWidget testStatusWidget(nullptr);
+    ControlPanelWidget controlPanelWidget(nullptr);
+
+    // QGridLayout* layout = new QGridLayout;
+    QVBoxLayout* layout = new QVBoxLayout;
+    layout->addWidget(&controlPanelWidget);
+    layout->addWidget(&channelWidget);
+    layout->addWidget(&testStatusWidget);
+    mainWidget.setLayout(layout);
+
+    mainWidget.show();
 
     spdlog::drop_all();
 
@@ -45,4 +70,11 @@ int main(int argc, char** argv)
     // TestControlOfPowerSupply();
     TestPSUController();
 #endif
+*/
+
+    QApplication app(argc, argv);
+    app.setStyle(QStyleFactory::create("Fusion"));
+    MainWindow window;
+    window.show();
+    return app.exec();
 }
